refactor discretisation structure and add tile coder
