+++ src/crl/cons/discretise/__init__.py
@@ -1,0 +1,2 @@
+from crl.cons.discretise.grid import run_test_episodes, build_tiling
+from crl.cons.discretise.tile_coder import build_tile_coding
